refactors: error handling & files.ts
